+++ okhttp/src/test/java/feign/okhttp/OkHttpClientTest.java
@@ -18,6 +18,7 @@ package feign.okhttp;
 import com.squareup.okhttp.mockwebserver.MockResponse;
 import com.squareup.okhttp.mockwebserver.rule.MockWebServerRule;
 
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -82,6 +83,7 @@ public class OkHttpClientTest {
   }
 
   @Test
+  @Ignore // TODO: Remove on OkHttp 2.5 https://github.com/square/okhttp/issues/1778
   public void patch() throws IOException, InterruptedException {
     server.enqueue(new MockResponse().setBody("foo"));
     server.enqueue(new MockResponse());